support scaling of vector graphics for logos, add xclim and cairosvg to dependencies, support more intuitive logo-handling, add examples

+++ figanos/_logo.py
@@ -19,17 +19,11 @@ _figanos_logo = Path(__file__).parent / "data" / "figanos_logo.png"
 class Logos:
     r"""Class for managing logos to be used in graphics.
 
-    Attributes
-    ----------
-    config : Path
-        The path to the folder where the logo configuration file is stored.
-    catalogue : Path
-        The path to the logo configuration file.
-    default : str
-        The path to the default logo.
-
     Methods
     -------
+    default
+        The path to the default logo.
+        can be temporarily set to a different logo via Logos().default = pathlib.Path().
     installed()
         Retrieves a list of installed logos.
     install_ouranos_logos(\*, permitted: bool = False)
@@ -39,16 +33,32 @@ class Logos:
         If no logos are already set, the first one will be set as the default.
     reload_config()
         Reloads the logo configuration from the YAML file.
-    """
 
-    default = None
+    Examples
+    --------
+    >>> from figanos import Logos
+    >>> logos = Logos()
+    >>> logos.default
+    PosixPath('/home/user/.config/figanos/logos/figanos_logo.png')
+    >>> logos.installed()
+    ['default', 'figanos_logo']
+    >>> logos.set_logo("path/to/logo.png", name="my_logo")
+    PosixPath('/home/user/.config/figanos/logos/my_logo.png')
+    >>> logos["my_logo"]
+    PosixPath('/home/user/.config/figanos/logos/my_logo.png')
+    >>> logos.default = "path/to/temporary/default/logo.png"
+    >>> logos
+    PosixPath("path/to/temporary/default/logo.png")
+    >>> logos.reload_config()
+    >>> logos.default
+    PosixPath('/home/user/.config/figanos/logos/figanos_logo.png')
+    """
 
     def __init__(self) -> None:
         """Constructor for the Logo class."""
-        self.config = (
-            Path(platformdirs.user_config_dir("figanos", ensure_exists=True)) / "logos"
-        )
-        self.catalogue = self.config / LOGO_CONFIG_FILE
+        self._config = None
+        self._catalogue = None
+        self._default = None
         self._logos = {}
         self.reload_config()
 
@@ -56,6 +66,31 @@ class Logos:
             warnings.warn(f"Setting default logo to {_figanos_logo}")
             self.set_logo(_figanos_logo)
             self.set_logo(_figanos_logo, name="default")
+
+    @property
+    def config(self) -> Path:
+        if self._config is None:
+            self._config = (
+                Path(platformdirs.user_config_dir("figanos", ensure_exists=True))
+                / "logos"
+            )
+        return self._config
+
+    @property
+    def catalogue(self) -> Path:
+        if self._catalogue is None:
+            self._catalogue = self.config / LOGO_CONFIG_FILE
+        return self._catalogue
+
+    @property
+    def default(self) -> str:
+        """The path to the default logo."""
+        return self._default
+
+    @default.setter
+    def default(self, value: Union[str, Path]):
+        """Set a default logo."""
+        self._default = value
 
     def _setup(self) -> None:
         if (
@@ -71,14 +106,23 @@ class Logos:
                 yaml.dump(dict(logos={}), f)
 
     def __str__(self) -> str:
-        return f"{self.__getitem__('default')}"
+        return f"{self._default}"
 
     def __repr__(self) -> str:
-        return f"{self._logos.items()}"
+        return f"{self._default}"
 
-    def __getitem__(self, name: str) -> Optional[str]:
-        """Retrieve a logo path by its name."""
-        return self._logos[name]
+    def __getitem__(self, args) -> Optional[str]:
+        """Retrieve a logo filepath by its name.
+
+        If it does not exist, it will be installed, with the filepath returned.
+        """
+        try:
+            return self._logos[args]
+        except (KeyError, TypeError):
+            if isinstance(args, tuple):
+                return self.set_logo(*args)
+            else:
+                return self.set_logo(args)
 
     def reload_config(self) -> None:
         """Reload the configuration from the YAML file."""
@@ -112,12 +156,15 @@ class Logos:
             _logo_mapping[name] = str(install_logo_path)
             self.catalogue.write_text(yaml.dump(dict(logos=_logo_mapping)))
             self.reload_config()
-            return self._logos[name]
+            if name != "default":
+                return self._logos[name]
+            else:
+                return self._default
 
         elif not logo_path.exists():
-            warnings.warn(f"Logo file {logo_path} not found. Not setting logo.")
+            warnings.warn(f"Logo file `{logo_path}` not found. Not setting logo.")
         elif not logo_path.is_file():
-            warnings.warn(f"Logo path {logo_path} is a folder. Not setting logo.")
+            warnings.warn(f"Logo path `{logo_path}` is a folder. Not setting logo.")
 
     def install_ouranos_logos(self, *, permitted: bool = False) -> None:
         """Fetches and installs the Ouranos logo.
@@ -133,12 +180,15 @@ class Logos:
             for orientation in ["horizontal", "vertical"]:
                 for colour in ["couleur", "blanc", "noir"]:
                     file = f"ouranos_logo_{orientation}_{colour}.png"
-                    logo_url = urllib.parse.urljoin(OURANOS_LOGOS_URL, file)
-                    try:
-                        urllib.request.urlretrieve(logo_url, self.config / file)
-                        self.set_logo(self.config / file)
-                    except Exception as e:
-                        logging.error(f"Error downloading or setting Ouranos logo: {e}")
+                    if not (self.config / file).exists():
+                        logo_url = urllib.parse.urljoin(OURANOS_LOGOS_URL, file)
+                        try:
+                            urllib.request.urlretrieve(logo_url, self.config / file)
+                            self.set_logo(self.config / file)
+                        except Exception as e:
+                            logging.error(
+                                f"Error downloading or setting Ouranos logo: {e}"
+                            )
 
             if Path(self.default).stem == "figanos_logo":
                 _default_ouranos_logo = (
